backend/workers/requirements.txt
celery==5.3.6
redis==5.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.6 
google-generativeai==0.3.2    
weaviate-client==3.26.2
pandas==2.1.3